% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igrf_map.R
\name{igrf_grid}
\alias{igrf_grid}
\title{International Geomagnetic Reference Field grid}
\usage{
igrf_grid(isv = 0, year, type = "spheroid", altitude, resolution = 5)
}
\arguments{
\item{isv}{main field (default = 0) or secular variation (1) data output}

\item{year}{year A.D. Must be greater than or equal to 1900.0 and
less than or equal to 2030. Warning message is given for dates
greater than 2025.}

\item{type}{"spheroid" (default) or "sphere" representation}

\item{altitude}{in km above the earth surface for a geodetic type,
or distance from the earth center for the spherical representation.}

\item{resolution}{spatial resolution of the output map in decimal degree
(default = 5).}
}
\value{
a data frame with components X,Y,Z,F, D, H and I for the main
geomagnetic field or the delta (dX etc) variants for the secular
variation with:
X = north component (nT) if isv = 0, nT/year if isv = 1
Y = east component (nT) if isv = 0, nT/year if isv = 1
Z = vertical component (nT) if isv = 0, nT/year if isv = 1
F = total intensity (nT) if isv = 0, rubbish if isv = 1

Reference:
International Geomagnetic Reference Field: the 13th generation
Alken, P., Thebault, E., Beggan, C.D. et al. International Geomagnetic
Reference Field: the thirteenth generation. Earth Planets Space 73, 49 (2021).
<https://doi.org/10.1186/s40623-020-01288-x>
}
\description{
Generates a grid (map) of the 13th Generation International Geomagnetic
Reference Field.
}
