% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igrf.R
\name{igrf}
\alias{igrf}
\title{International Geomagnetic Reference Field}
\usage{
igrf(isv = 0, year, type = "spheroid", altitude, latitude, longitude)
}
\arguments{
\item{isv}{main field (default = 0) or secular variation (1) data output}

\item{year}{year A.D. Must be greater than or equal to 1900.0 and
less than or equal to 2030. Warning message is given for dates
greater than 2025.}

\item{type}{"spheroid" (default) or "sphere" representation}

\item{altitude}{in km above the earth surface for a geodetic type,
or distance from the earth center for the spherical representation.}

\item{latitude}{latitude in decimal degrees}

\item{longitude}{longitude in decimal degrees}
}
\value{
a data frame with components X,Y,Z,F, D, H and I for the main
geomagnetic field or the delta (dX etc) variants for the secular
variation with:
X = north component (nT) if isv = 0, nT/year if isv = 1
Y = east component (nT) if isv = 0, nT/year if isv = 1
Z = vertical component (nT) if isv = 0, nT/year if isv = 1
F = total intensity (nT) if isv = 0, rubbish if isv = 1

Reference:
International Geomagnetic Reference Field: the 13th generation
Alken, P., Thebault, E., Beggan, C.D. et al. International Geomagnetic
Reference Field: the thirteenth generation. Earth Planets Space 73, 49 (2021).
<https://doi.org/10.1186/s40623-020-01288-x>
}
\description{
The 13th Generation International Geomagnetic Reference Field.
This is an implementation of the Fortran code provided on
the NOAA website <https://www.ngdc.noaa.gov/IAGA/vmod/igrf.html>
}
\details{
The main code driving the model output is taken from the original Fortran
model published by Alken et al. 2021. Outputs have been verified to
correspond with the original code with the exception that values in this
implementation are not rounded before output (as in the original model).
For all intents and purposes the data can be considered equivalent.

For full model details we refer Alken et al. 2021, and the project website
where both the original code and a brief model description can be found.

The authors of both the R package and the original Fortran code
take no responsibility regarding the use of these data within
a professional context.
}
